fix(pluginui): coerce whole float badge counts to int so shells render 8 not 8.000000
buildNavItems stored the raw JSON badge count (float64 8.0); the shell templates
printed 8.000000. wholeNumber() maps whole floats to int64, passing non-whole and
non-numeric values through. Test updated to assert int64(5).

+++ internal/platform/pluginui/router_test.go
@@ -227,7 +227,7 @@ func TestRegisterUIRoutes(t *testing.T) {
 			t.Error("Home should be active on /")
 		}
 		// Second item (Items) should have a badge count.
-		if navItems[1]["badge_count"] != float64(5) {
+		if navItems[1]["badge_count"] != int64(5) {
 			t.Errorf("badge_count = %v, want 5", navItems[1]["badge_count"])
 		}
 		// ui_nav must be a map whose "position" the shell templates can resolve
